DjangoUserService/pyproject.toml
[project]
name = "DjangoUserService"
version = "0.1.0"
description = "Django User Service"
authors = [
    { name = "RMA-MUN", email = "n3032747608@163.com" },
]
requires-python = ">=3.10"
dependencies = [
    "django==5.2.6",
    "PyMySQL==1.1.0",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.3.1",
    "django-cors-headers~=4.3.0",
    "celery==5.4.0",
    "redis==5.0.1",
    "django-redis==5.4.0",
    "python-dotenv==1.0.0",
    "uvicorn>=0.40",
    "asgiref>=3.11",
    "setuptools<60",
    "drf-yasg>=1.21.15",
    "django-shortuuidfield>=0.1.3",
    "pillow>=12.1.1",
    "mysqlclient>=2.2.8",
    "mysql-connector-python>=9.6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["DjangoUserService", "apps"]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"
